src/main/java/ua/abdulaiev/hw26/dao/AlbumDao.java
package ua.abdulaiev.hw26.dao;

import org.hibernate.Session;
import org.hibernate.Transaction;
import ua.abdulaiev.hw26.enums.AlbumNames;
import ua.abdulaiev.hw26.enums.ArtistNames;
import ua.abdulaiev.hw26.enums.Tracks;
import ua.abdulaiev.hw26.models.Album;
import ua.abdulaiev.hw26.models.Artist;
import ua.abdulaiev.hw26.utils.HibernateUtils;

import java.util.ArrayList;
import java.util.List;
import java.util.Random;
import java.util.concurrent.ThreadLocalRandom;


public class AlbumDao {
    public void save(Album album) {
        Transaction transaction = null;
        try (Session session = HibernateUtils.getSessionFactory().openSession()) {
            transaction = session.beginTransaction();
            session.save(album);
            transaction.commit();
        } catch (Exception e) {
            if (transaction != null) {
                transaction.rollback();
            }
            e.printStackTrace();
        }
    }

    public void update(Album album) {
        Transaction transaction = null;
        try (Session session = HibernateUtils.getSessionFactory().openSession()) {
            transaction = session.beginTransaction();
            session.update(album);
            transaction.commit();
        } catch (Exception e) {
            if (transaction != null) {
                transaction.rollback();
            }
            e.printStackTrace();
        }
    }

    public void delete(int id) {
        Transaction transaction = null;
        try (Session session = HibernateUtils.getSessionFactory().openSession()) {
            transaction = session.beginTransaction();
            Album album = session.get(Album.class, id);
            if (album != null) {
                session.delete(album);
            }
            transaction.commit();
        } catch (Exception e) {
            System.out.println(e.getMessage());
            if (transaction != null) {
                transaction.rollback();
            }
        }
    }

    public Album getById(int id) {
        Transaction transaction = null;
        Album album = null;
        try (Session session = HibernateUtils.getSessionFactory().openSession()) {
            transaction = session.beginTransaction();
            album = session.get(Album.class, id);
            transaction.commit();
        } catch (Exception e) {
            if (transaction != null) {
                transaction.rollback();
            }
            e.printStackTrace();
        }
        return album;
    }

    public List<Album> getAllAlbum() {
        Transaction transaction = null;
        List<Album> albums = null;
        try (Session session = HibernateUtils.getSessionFactory().openSession()) {
            transaction = session.beginTransaction();
            albums = session.createQuery(
                    "SELECT a FROM Album a", Album.class)
                    .list();
            transaction.commit();
        } catch (Exception e) {
            if (transaction != null) {
                transaction.rollback();
            }
            e.printStackTrace();
        }
        return albums;
    }
}